Don't take a goTo() right after a drag
The goTo() was being called because a child was being "clicked" after the drag was complete, meaning that the drag goTo() was immediately overridden by a "click" goTo().

No one clicking again within 300ms will likely want that result, so just debounce it and ignore any click that comes in sooner.